Move PC hit-point overage handling to base item class
Besides ABC items, heritage and feats also need to be handled
